requirements.txt
rapidocr-onnxruntime
rapidfuzz
python-dotenv
